+++ src/main/java/co/geomati/olakease/api/DeveloperResource.java
@@ -21,21 +21,21 @@ public class DeveloperResource extends
 	@GET
 	@Produces(MediaType.APPLICATION_JSON)
 	@Override
-	public Developer asJSON(@PathParam("developerId") long id) {
+	public Developer asJSON(@PathParam("developerId") int id) {
 		return super.asJSON(id);
 	}
 
 	@PUT
 	@Produces(MediaType.APPLICATION_JSON)
 	@Override
-	public Developer put(@PathParam("developerId") long id,
+	public Developer put(@PathParam("developerId") int id,
 			Developer modifications) {
 		return super.put(id, modifications);
 	}
 
 	@Override
-	protected Developer mergeModifications(EntityManager entityManager,
-			long id, Developer modifications) {
+	protected Developer mergeModifications(EntityManager entityManager, int id,
+			Developer modifications) {
 		modifications.setId(id);
 		return modifications;
 	}
